+++ UnitySDK/Packages/com.happs.sdk/Runtime/Mobile/Models/MobileCreatePaymentRequest.cs
@@ -1,0 +1,11 @@
+namespace HAppsSDK
+{
+	public sealed class MobileCreatePaymentRequest
+	{
+		public string ProductId;
+		public decimal Price;
+		public string Currency;
+		public string Description;
+		public string RequestId;
+	}
+}