+++ UnitySDK/Packages/com.happs.sdk/Runtime/Mobile/Models/MobileCreatePaymentResult.cs
@@ -1,0 +1,8 @@
+namespace HAppsSDK
+{
+	public sealed class MobileCreatePaymentResult
+	{
+		public string OrderId;
+		public string PaymentUrl;
+	}
+}