+++ UnitySDK/Packages/com.happs.sdk/Runtime/Models/HAppsMessage.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace HAppsSDK
+{
+	[Serializable]
+	public class HAppsMessage
+	{
+		public string type;
+
+		public InitData initData;
+		public UserData userData;
+		public SignatureData signatureData;
+		public PaymentData paymentData;
+		public AuthPopupData authPopupData;
+	}
+}