+++ UnitySDK/Packages/com.happs.sdk/Runtime/Models/InitData.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace HAppsSDK
+{
+	[Serializable]
+	public class InitData
+	{
+		public bool ready;
+		public bool fromPlatform;
+		public bool isMobileWeb;
+	}
+}