+++ UnitySDK/Packages/com.happs.sdk/Runtime/Models/SignatureData.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace HAppsSDK
+{
+	[Serializable]
+	public class SignatureData
+	{
+		public string signature;
+	}
+}