+++ UnitySDK/Packages/com.happs.sdk/Runtime/Models/UserData.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace HAppsSDK
+{
+	[Serializable]
+	public class UserData
+	{
+		public string userId;
+		public string userName;
+		public bool verified;
+
+		public override string ToString()
+		{
+			return $"{userName} ({userId})";
+		}
+	}
+}